Enumerate value changes in yaml file
