+++ camel-upgrade-recipes/src/main/resources/META-INF/rewrite/latest.yaml
@@ -29,7 +29,7 @@ recipeList:
   - org.apache.camel.upgrade.camel40.CamelMigrationRecipe
   - org.apache.camel.upgrade.UpgradeToJava17
   - org.apache.camel.upgrade.JavaVersion17
-  - org.apache.camel.upgrade.CamelPropertiesAndYamlUpdate
+  - org.apache.camel.upgrade.UpdatePropertiesAndYamlKeys
   - org.openrewrite.maven.UpgradeDependencyVersion:
       groupId: 'org.apache.camel'
       artifactId: '*'
